% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropped-average.R
\name{avg_w_drops}
\alias{avg_w_drops}
\title{Average of grades after dropping lowest scores}
\usage{
avg_w_drops(scores, ndrop)
}
\arguments{
\item{scores}{A numeric vector giving the raw scores; values should be in
\eqn{[0, 1]} or \eqn{[0, 100]}}

\item{ndrop}{Number of assignments to drop}
}
\value{
The grades' average after dropping the relevant number of assignments
}
\description{
If the number \code{ndrop} of assignments to drop is not a whole number, a
weighted average is given, where the \code{length(scores) - ceiling(ndrop)}
highest scores are weighted normally, but the next highest score is given a
weight of \code{r} (and all lower scores receive a weight of 0).
}
