% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bates_watts_D.R
\name{bates_watts_D}
\alias{bates_watts_D}
\title{Computes Bates and Watts D matrix}
\usage{
bates_watts_D(f, x, eps = 1e-04, d = 1e-04, r = 4, v = 2,
  zero_tol = sqrt(.Machine$double.eps/7e-07))
}
\arguments{
\item{f}{Fucntion}

\item{x}{Point}

\item{eps}{See \code{numderiv::genD}}

\item{d}{See \code{numderiv::genD}}

\item{r}{See \code{numderiv::genD}}

\item{v}{See \code{numderiv::genD}}

\item{zero_tol}{See \code{numderiv::genD}}
}
\description{
Computes Bates and Watts D matrix
}

