% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{calculateOverallAverage}
\alias{calculateOverallAverage}
\title{The averaging function.}
\usage{
calculateOverallAverage(
  table,
  categories,
  studentNameCol,
  drop = NULL,
  traceCalcsForStudent = NULL
)
}
\arguments{
\item{table}{a `data.frame` of student grades.}

\item{categories}{a named list of named lists. See below for more information.}

\item{studentNameCol}{a name of a column with student names.}

\item{drop}{a list of named lists. See below for more information.}

\item{traceCalcsForStudent}{Character scalar giving a student name, or NULL. If provided, additional intermediate calculations are printed for that student only (useful for debugging).}
}
\description{
This function calculates each student's course average based on all of their assignments.
}
\examples{
\donttest{
exampleDF <- data.frame(student = c('a','b','c'),
                        test1 = c(50,50,NA),
                        test1LateAdjustment = c(.99,1,1),
                        hw1=c(45,44,43),
                        hw1LateAdjustment = c(1,1,1),
                        hw2=c(50,49,48),
                        hw2LateAdjustment = c(1,1,1))

testCols <- data.frame(gradeNames = 'test1', 
                       lateNames = 'test1LateAdjustment', 
                       maxScores = 50)
testCategory <- list(colNames=testCols,
                     catWeight=.6)
hwCols <- data.frame(gradeNames = c('hw1','hw2'), 
                     lateNames = c('hw1LateAdjustment','hw2LateAdjustment'),
                     maxScores = c(50,50))
homeworkCategory <- list(colNames=hwCols,
                         catWeight=.4)

cats <- list(tests=testCategory, homeworks=homeworkCategory)
myDrops = list(numToDrop=1, cats='homeworks')
calculateOverallAverage(exampleDF, cats, 'student', myDrops)
calculateOverallAverage(exampleDF, cats, 'student', myDrops, traceCalcsForStudent = 'a')
}
}
\keyword{calculateOverallAverage}
