% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createKey.R
\name{createKey}
\alias{createKey}
\title{Create an Answer Key}
\usage{
createKey(key_name = "homework", vars = NA)
}
\arguments{
\item{key_name}{Full path to the assignment name.  ("_KEY.Rdata" will be appended)}

\item{vars}{Names of the variables where answers are stored.  (default = all variables in current workspace)}
}
\value{
An .Rdata file containing the list 'answers' with solutions.
}
\description{
Create an Answer Key
}
\examples{
\dontrun{
x <- 9
y <- 1:10
z <- rep(0,5)

# Save all variables as 'test_KEY.Rdata'
createKey("test")

# Save only x & y as 'test_KEY.Rdata'
createKey("test", vars = c("x", "y"))
}

}
\seealso{
\code{\link{grade}}
}

