% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.R
\name{curve_grades}
\alias{curve_grades}
\title{Curve grades}
\usage{
curve_grades(
  scores,
  method = "linear",
  x0 = mean(scores),
  x1 = quantile(scores, probs = 0.95),
  y0 = 85,
  y1 = 95,
  target_mean = 85,
  target_sd = 5
)
}
\arguments{
\item{scores}{A numeric vector giving the raw scores; values should be in
\eqn{[0, 1]}}

\item{method}{A character vector, one of "linear" or "normal" (see details);
default: "linear"}

\item{x0, x1}{If \code{method == "linear"}, two raw scores to serve as anchors;
defaults are \code{mean(scores)} and \code{quantile(scores, probs = 0.95)}}

\item{y0, y1}{If \code{method == "linear"}, what \code{x0} and \code{x1} should map onto;
defaults are 85 and 95}

\item{target_mean}{If \code{method == "normal"}, the mean of the resulting
normal distribution; default: 85}

\item{target_sd}{If \code{method == "normal"}, the standard deviation of the
resulting normal distribution; default: 5}
}
\value{
A numeric vector giving the curved scores
}
\description{
Both a linear curving tool and a normal curving tool are provided
(see details)
}
\details{
Two grade curving methods are implemented:
\enumerate{
\item Linear curve: Takes two scores, \code{x0} and \code{x1}, and ensures those map onto
two scores \code{y0} and \code{y1}, and linearly scales the remaining scores. For
example, you may want to ensure the mean was a certain value like 85
and the 95\% percentile grade was actually 95 (the default behavior),
or that the mean was 85 and the lowest score was 50, etc.
\item Normal curve: Matches the empirical quantiles of your students' scores
with the theoretical quantiles of a normal distribution with the given
\code{target_mean} and \code{target_sd} truncated to \eqn{[0, 100]}.
}
}
