% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{findBadEncodingFiles}
\alias{findBadEncodingFiles}
\title{This function finds unreadable files.}
\usage{
findBadEncodingFiles(submission_dir)
}
\arguments{
\item{submission_dir}{where the assignments are located}
}
\description{
A function that finds student submissions with poorly encoded characters
}
\examples{
\donttest{
# change paths to *your* paths
submissions <- "extdata/assignment1_submissions/"
findBadEncodingFiles(submissions) # perhaps ask these students to resubmit
}
}
\keyword{calcGrades}
\keyword{findBadEncodingFiles}
