% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{findGlobalPaths}
\alias{findGlobalPaths}
\title{This function finds files with global file paths.}
\usage{
findGlobalPaths(submission_dir)
}
\arguments{
\item{submission_dir}{where the assignments are located}
}
\description{
A function that finds student submissions that refer to machine-specific file paths
}
\examples{
\donttest{
# change paths to *your* paths
submissions <- "extdata/assignment1_submissions/"
findGlobalPaths(submissions) # perhaps ask these students to resubmit
}
}
\keyword{calcGrades}
\keyword{findGlobalPaths}
