% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{getPrettyReport}
\alias{getPrettyReport}
\title{The function prints a formatted test script analysis.}
\usage{
getPrettyReport(report, script_path)
}
\arguments{
\item{report}{a list object that is produced from the getTestScriptReport function}

\item{script_path}{a string representing the relative path to the test script}
}
\description{
This function is used inside the getTestScriptReport function to produce a formatted report as output.
}
\keyword{internal}
