% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Grades.R
\name{grade_plot}
\alias{grade_plot}
\title{Calculate Grades}
\usage{
grade_plot(pdfname, score, cuts)
}
\arguments{
\item{pdfname}{name of pdf file}

\item{score}{matrix of class scores}

\item{cuts}{cuttoff points for letter grades}
}
\value{

}
\description{
Calculate Grades
}
