% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradeR-package.R
\docType{package}
\name{gradeR-package}
\alias{gradeR}
\alias{gradeR-package}
\title{gradeR: Help grade R script assignment submissions!}
\description{
This is a single function R package that helps grade students' R script assignment submissions. This function finds every .R file in a directory, runs it, and tests variables are what they are expected to be.
}
\author{
\strong{Maintainer}: Taylor Brown \email{trb5me@virginia.edu}

}
\keyword{internal}
