% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gradient.R
\name{gradient}
\alias{gradient}
\title{Compute gradient of a function}
\usage{
gradient(f, x, eps = 1e-04, d = 1e-04, r = 4,
  zero_tol = sqrt(.Machine$double.eps/7e-07))
}
\arguments{
\item{f}{Function}

\item{x}{Point to compute gradient at}

\item{eps}{See \code{numDeriv::grad}}

\item{d}{see \code{numDeriv::grad}}

\item{r}{see \code{numDeriv::grad}}

\item{zero_tol}{see \code{numDeriv::grad}}
}
\description{
Compute gradient of a function
}

