% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{groups2student}
\alias{groups2student}
\title{Randomly Assign Groups to Student Discussants}
\usage{
groups2student(g = 9, n = 6, w = 14, v = 40)
}
\arguments{
\item{g}{number of groups}

\item{n}{number of people per group}

\item{w}{number of weeks}

\item{v}{number of students}
}
\value{

}
\description{
Randomly Assign Groups to Student Discussants
}
