% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use.R
\name{Hessian}
\alias{Hessian}
\title{Compute Hessian of a function}
\usage{
Hessian(f, x, ...)
}
\arguments{
\item{f}{Function}

\item{x}{Point to compute Hessian at}

\item{...}{Additional arguments passed to \code{\link{bates_watts_D}}.}
}
\description{
Compute Hessian of a function
}

