% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{i.n.cases.correct.answers}
\alias{i.n.cases.correct.answers}
\title{i.n.cases.correct.answers}
\usage{
i.n.cases.correct.answers(n.cases.correct.answers, q.answers, i.cases)
}
\arguments{
\item{i.answers}{}
}
\value{

}
\description{
i.n.cases.correct.answers
}
