% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{i.test.fun}
\alias{i.test.fun}
\title{i.test.fun}
\usage{
i.test.fun(i.test, f.out, n.test, n.question, bank = TRUE, Qend = TRUE,
  latex.dir.out = "latexOut", do.randomize.questions = TRUE,
  do.randomize.answers = TRUE, list.in, l.def, verbose = FALSE)
}
\arguments{
\item{do.randomize.questions}{}

\item{verbose}{}
}
\value{

}
\description{
i.test.fun
}
