% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{latexout}
\alias{latexout}
\title{latexout}
\usage{
latexout(my.tex.file, f.temp.tex, bank, exam.class, str.pattern.end.mchoice,
  my.last.part, qtext, Qend = TRUE)
}
\arguments{
\item{Qend=TRUE}{}
}
\value{

}
\description{
latexout
}
