% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Grades.R
\name{letter_calc}
\alias{letter_calc}
\title{Calculate Letter Grades}
\usage{
letter_calc(scores, cuts, percent = FALSE)
}
\arguments{
\item{weight_df}{number of groups}

\item{n}{number of people per group}

\item{w}{number of weeks}

\item{v}{number of students}
}
\value{

}
\description{
Calculate Letter Grades
}
