% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{my.test.comment}
\alias{my.test.comment}
\title{helper commands to rte.analyze.tex.file to eliminate coded out lines}
\usage{
my.test.comment(str.in)
}
\arguments{
\item{str.in}{}
}
\value{
logical TRUE if line coded out
}
\description{
helper commands to rte.analyze.tex.file to eliminate coded out lines
}
