% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{parseTestResults}
\alias{parseTestResults}
\title{Parse test results for a single student.}
\usage{
parseTestResults(test_results, number_questions)
}
\arguments{
\item{test_results}{the ListReporter results object}

\item{number_questions}{the number of questions/tests}
}
\value{
A numeric vector of scores (0 or 1 for each question)
}
\description{
Internal helper function that evaluates test results and returns scores.
}
\keyword{internal}
