% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{q.answers.cases.fun}
\alias{q.answers.cases.fun}
\title{q.answers.cases.fun}
\usage{
q.answers.cases.fun(case.now, q.answers, str.pattern.correct,
  str.pattern.choice)
}
\arguments{
\item{str.pattern.choice}{}
}
\value{

}
\description{
q.answers.cases.fun
}
