% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{rand_fun}
\alias{rand_fun}
\title{Batch Latex Export}
\usage{
rand_fun(lfile, latex.dir.in, latex.dir.out, n.test = 1,
  do.randomize.questions = TRUE, do.randomize.answers = TRUE)
}
\arguments{
\item{do.randomize.answers}{}
}
\value{

}
\description{
Batch Latex Export
}
