% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{random.matrix}
\alias{random.matrix}
\alias{random.matrix.no.dup}
\title{Randomly Assign Student Discussants}
\usage{
random.matrix(p, v)

random.matrix.no.dup(p, v)
}
\arguments{
\item{p}{number of times each person presents}

\item{v}{number of students}
}
\value{

}
\description{
Randomly Assign Student Discussants

Randomly Assign Student Discussants w/o duplicates
}
