% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{rte.analyze.tex.file}
\alias{rte.analyze.tex.file}
\title{import and analyze a latex file}
\usage{
rte.analyze.tex.file(f.in, bank = TRUE)
}
\arguments{
\item{f.in}{character string for input file}
}
\value{
exam
}
\description{
import and analyze a latex file
}
