% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLatexExam.R
\name{rte.build.rdn.test}
\alias{rte.build.rdn.test}
\title{rte.build.rdn.test}
\usage{
rte.build.rdn.test(list.in, f.out, n.test, n.question, bank = TRUE,
  latex.dir.out = "latexOut", do.randomize.questions = TRUE,
  do.randomize.answers = TRUE)
}
\arguments{
\item{do.randomize.answers=TRUE}{}
}
\value{

}
\description{
rte.build.rdn.test
}
