% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Grades.R
\name{score_calc}
\alias{score_calc}
\title{Calculate Grades}
\usage{
score_calc(weight_df, verbose = TRUE)
}
\arguments{
\item{weight_df}{number of groups}

\item{verbose}{print output}
}
\value{

}
\description{
Calculate Grades
}
