% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use.R
\name{slope_info}
\alias{slope_info}
\title{Compute information about slopes}
\usage{
slope_info(f, x, ...)
}
\arguments{
\item{f}{Function}

\item{x}{Point}

\item{...}{Additional arguments to \code{\link{bates_watts_D}}.}
}
\description{
Compute information about slopes
}
\seealso{
\code{\link{taylor2}}
}

