% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{students2weeks.format}
\alias{students2weeks.format}
\title{Formatting Table for Export}
\usage{
students2weeks.format(discussant_table, students, write_file = NA)
}
\arguments{
\item{discussant_table}{table to format}

\item{students}{matrix of students}

\item{write_file}{write table to csv? default NA}
}
\value{
a vector
}
\description{
Formatting Table for Export
}
