% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{students2weeks.matrix}
\alias{students2weeks.matrix}
\alias{students2weeks.matrix.no.dup}
\title{Randomly Assign Student Discussants to Weeks}
\usage{
students2weeks.matrix(students, w = 14, p = NA, v = NA)

students2weeks.matrix.no.dup(students, w = 14, p = NA, v = NA)
}
\arguments{
\item{students}{matrix of students}

\item{w}{number of weeks}

\item{p}{number of times each person presents}

\item{v}{number of students}
}
\value{

}
\description{
Randomly Assign Student Discussants to Weeks

Randomly Assign Student Discussants to Weeks without duplicates
}
