% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{students2weeks.matrix2}
\alias{students2weeks.matrix2}
\alias{students2weeks.matrix2.no.dup}
\title{Randomly Assign Student Discussants}
\usage{
students2weeks.matrix2(v, w, p = NA, a = 2)

students2weeks.matrix2.no.dup(v, w, p = NA, a = 2)
}
\arguments{
\item{v}{number of students}

\item{w}{number of weeks}

\item{p}{number of times each person presents}

\item{a}{}
}
\value{
a vector
}
\description{
Randomly Assign Student Discussants

Randomly Assign Student Discussants
}
