% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use.R
\name{taylor2}
\alias{taylor2}
\title{Approximate a function}
\usage{
taylor2(obj)
}
\arguments{
\item{obj}{Output from \code{\link{slope_info}}}
}
\description{
Create a 2nd order Taylor series (Laplace's method) approximation
to a function.
}

