% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{calcGrades}
\alias{calcGrades}
\title{A grading function!}
\usage{
calcGrades(submission_dir, your_test_file)
}
\arguments{
\item{submission_dir}{where the assignments are located}

\item{your_test_file}{the path to your testthat test file (e.g. grade_hw1.R)}
}
\description{
This function grades a bunch of R script assignments
}
\examples{
\donttest{
# change paths to *your* paths
submissions <- "extdata/assignment1_submissions/"
my_test_file <- system.file("extdata", "grade_hw1.R", package = "gradeR")
results <- calcGrades(submissions, my_test_file)
}
}
\keyword{calcGrades}
