% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{calcGrades}
\alias{calcGrades}
\title{The grading function.}
\usage{
calcGrades(
  submission_dir,
  your_test_file,
  suppress_warnings = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{submission_dir}{where the assignments are located}

\item{your_test_file}{the path to your testthat test file (e.g. grade_hw1.R)}

\item{suppress_warnings}{warning handlers prevent code from being run after they catch something. Suppress this behavior by setting this argument to TRUE.}

\item{verbose}{set to true if you want to print the name of the file as it's being ran}
}
\value{
a data.frame of all the scores for each student
}
\description{
This function grades a bunch of R script assignments
}
\examples{
\donttest{
# change paths to *your* paths
submissions <- "extdata/example/assignment1_submissions/"
my_test_file <- system.file("extdata/example", "grade_hw1.R", package = "gradeR")
results <- calcGrades(submissions, my_test_file)
}
}
\keyword{calcGrades}
