% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{calcGradesForGradescope}
\alias{calcGradesForGradescope}
\title{The grading function for Gradescope.}
\usage{
calcGradesForGradescope(
  submission_file,
  test_file,
  which_results = "gradescope"
)
}
\arguments{
\item{submission_file}{the name of the assignment submission file (e.g. "hw1.r")}

\item{test_file}{the name of the .r file with test_that tests (e.g. "hw1_tests.R")}

\item{which_results}{Choose either "testing" or "gradescope" If equal to "gradescope" then the json file is written out to the directory that Gradescope expects. Otherwise, results.json is written to your current working directory.}
}
\description{
This function grades one R script assignment submission and writes results out to a properly-formatted json file for Gradescope.
}
\keyword{Gradescope}
\keyword{calcGradesForGradescope}
