% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{getTestScriptReport}
\alias{getTestScriptReport}
\title{The function for analyzing and summarizing R test scripts.}
\usage{
getTestScriptReport(script_path)
}
\arguments{
\item{test_file}{the name of the .r file with test_that tests (e.g. "hw1_tests.R")}
}
\description{
This function scans a given test script and summarizes the number of tests and test criteria in the script, as well as point values.
}
\keyword{getTestScriptReport}
