% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Grade based on R workspace equivalence}
\usage{
grade(assignment, key, grade_for_completion = FALSE)
}
\arguments{
\item{assignment}{Full path to assignment (.Rdata) (character).  Can be a vector.}

\item{key}{Full path to key (.Rdata) (character).}

\item{grade_for_completion}{Should grading be based on completion or accuracy (default = FALSE).}
}
\value{
Data.frame with the file names and corresponding grades.
}
\description{
Grade based on R workspace equivalence
}
\examples{
\dontrun{
my_key <- file.path(getwd(), "Lesson1_KEY.Rdata")

student_names <- c("Anthony", "Shannon", "Avery")
students <- file.path(getwd(), paste0("Lesson1_", student_names, ".Rdata"))

grade(students, my_key, TRUE)
grade(students, my_key, FALSE)
}

}
\seealso{
\code{\link{createKey}}
}

