% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{gradeR}
\alias{gradeR}
\title{A grading function!}
\usage{
gradeR(submission_dir, your_test_file)
}
\arguments{
\item{submission_dir}{where the assignments are located}

\item{your_test_file}{the path to your testthat test file (e.g. grade_hw1.R)}
}
\description{
This function grades a bunch of R script assignments
}
\examples{
submissions <- "~/gradeR/assignment1_submissions/"
my_test_file <- "~/gradeR/my_tests/grade_hw1.R"
results <- gradeR(submissions, my_test_file)
}
\keyword{cats}
