% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.r
\name{processTestAssertions}
\alias{processTestAssertions}
\title{Process assertions for a single test.}
\usage{
processTestAssertions(assertion_results)
}
\arguments{
\item{assertion_results}{list of assertion results from testthat}
}
\value{
A list with components: score, max_score, and output
}
\description{
Internal helper function that processes all assertions for a test and
computes scores and messages.
}
\keyword{internal}
