% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeachingFunctions_Assignments.R
\name{students2weeks.print}
\alias{students2weeks.print}
\title{Write A Latex Table for Each Week}
\usage{
students2weeks.print(discussants, w = 13, rdir)
}
\arguments{
\item{discussants}{list/matrix of discussants}

\item{w}{number of weeks}

\item{rdir}{directory to write discussants to}
}
\value{
a vector
}
\description{
Write A Latex Table for Each Week
}
